#1 Create statically linked binary
If the binary does not link all functions statically, the binary will call
glibc functions for net or user functionality (which is not available in
distroless or alpine containers). This commit modifies the go linker flags in
a way that only a statically binary will be produced.

The advantage of being independent seemingly does not show in the size of the
binary because both binaries (statically or dynamically) show up as 46 MB.