#4 Replace YAML namespaces with namespace from app config
The dogu operator comes with its own namespace properties inside the YAML file.
That leads to problems because the dogu operator should be installed inside the
desired target namespace. Since these two do not match, the K8s API throws an
error.

This commit replaces the namespace with a regexp despite having YAML. For YAML
we most probably need a proper struct which we don't know beforehand.