#4 Run setup only within a single namespace
This commit changes all previous behavior in which the setup was expected to run
in a different namespace than the target namespace. The reason is to use a less
demanding environment.

In the long run, the currently necessary action can be fulfilled by a new
operator, so that the administrator has even less tasks to do before the
installation.

This commit also removes a lot of cluster role permissions in favor of a single
permission for creating/updating the dogu CRD that comes in via the dogu
operator.