Apply controller reference and namespace for namespaced resourced only
For some resources an empty namespace field may lead to an error when applied to
the K8s API. This commit generally sets the resource's namespace field because
otherwise this may lead to API errors during patching.

Additionally, the controller reference will only be set to namespaced resources
because setting them to cluster-wide resources is not allowed for obvious
reasons.